[refactor]ArticleRepositoryを作成し、ユーザー投稿の登録処理をリファクタリング
