--- laravel/app/Http/Requests/Article/ArticleRepository.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Repositories\User;
-
-class ArticleRepository implements ArticleRepositoryInterface
-{
-
-}
