--- laravel/app/Http/Requests/Article/ArticleRepositoryInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Repositories\Article;
-
-interface ArticleRepositoryInterface
-{
-
-}
